modification_hooks: potential speed up
For certain files, the modification hooks are one of the big time wasters.
This patch sort of combines the strlen() and the strncmp() into one
function.

I implemented this late at night and had to leave in a hurry when my baby
woke up and started crying so I can't remember what I was testing and
don't have before and after figures to say if it really helps.  It might
not, but it felt good and I hate to throw away work so I'm committing it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>