select: simplify handling of select(x, 0, x) --> 0
This simplification is already handled but explicitly kills it's 2
operands while this is done automatically when killing the instruction.

Also, it uses replace_with_value(0) which needs to recreate the
pseudo for 0 while it's already available via its operands.

So, changes to use replace_with_pseudo() and without the unneeded kills.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>