parse_call_math: fix a segfault
Sometimes the call math is "r $2" which means that the $2 argument is a
function pointer and the return is whatever the function pointer returns.
But then say that the function pointer returns "$1->len".  So now we need
to know the argument that was passed to the pointer.  But we don't have
that.

Instead what was happening was that Smatch was treating the pointer as a
call and dereferencing something random and crashing.

So check that the "call" really is a call.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>