param_used: ignore some states used to fill the DB
I have a certain kind of paranoia about the PARAM_USED stuff taking
too much space in the DB.  Things from storing stuff in the DB is not
supposed to be counted as used.

It turns out this code hasn't actually been enabled until now so it isn't
tested but it should be correct.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>