param_set: handle certain struct assignments
The problem here is that if a function does a struct assignment and
sets a struct member that should be recorded.  In other words

	*param = foo;

results in param->a and param->b being set.

The problem is that just recording all assignments bloated the DB so much
that it was unworkable.  The fix for that is maybe to mark a whole range
of data as unknown.  But that's not the fix I implemented here.  My fix
is more of a hack.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>