fix trivial_phi() when the target is before the single value
A phi-node is called 'trivial' if it only reference itself and a
single other value. In this case the only possible value for the
phi-node is this single other value which can thus replace the
phi-node.

However, the current code get this slightly wrong when the first
referenced value is itself and not the other value.

Fix this by moving up the test checking if it references itself.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>