kill parent's dead stores too
kill_dominated_stores() identify and remove dead stores
(stores unneeded because the same location is overwritten later
by another store) only when both stores are in the same basic block.

Slightly improve this by also handling the case when the dead store
is in a parent BB of the "live" store.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>