readl_infinite_loops: delete this check
The truth is that hardware which can recover from a PCI error is really
rare and generally if the hardware is that broken you should just replace
it.  It doesn't make a lot of sense to try work around it.

This check worked but the warnings weren't something that most people
thought was worth fixing.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>