check_deref: silence false positives
Smatch had a bug because it complained about:

    drivers/gpu/drm/drm_color_mgmt.c:307 drm_crtc_legacy_gamma_set()
    error: potential null dereference 'blob'.  (drm_property_create_blob returns null)

The problem is that drm_crtc_legacy_gamma_set() returns error pointers and
not NULL.  This didn't trigger if the database was built (which is how I
normally test it) but it does trigger for almost everyone else.

My new patch effectively silences this check unless the database is built
which is slightly unfortunate but probably the solution going forward.

Reported-by: Philip Li <philip.li@intel.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>