ptrlist: avoid mixing reverse and non-reverse macros
The macros used to iterate the ptrlists exist in two kinds:
those to iterate forward direction and those to iterate in the
reverse direction.

Those macros must be used in pair: one for the top of the loop and
one at the end of the loop. However, if we mix them, for example like:
	FOR_EACH_PTR(list, var) {
		...
	} FOR_EACH_PTR_REVERSE(var);

things will still work for lists with a single block (most of them)
but will behave strangely and of course wrongly when reaching the
next block.

So, to avoid future debugging fun, add a unused variable, discarded
at compile time, but with distinct prefix for each direction. This way,
mixing the macros will create a warning at compile time.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>