states: use the state from the fake assignment
There was no reason for this, it just seemed like an good idea so I
did it.

The problem was that inline functions get parsed more than once so it led
to double frees.  So I added a check for that.  Otherwise, I didn't notice
any difference in the output.  But I'm only looking at bad output so maybe
it made it better and I didn't notice.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>