points_to_user_data: Make copy_from_user() set point to user data
Say we have a pointer "p" and we copy user data to it then it points to
user data.  And if we do p++ then that's points to user data as well.
If "p" is a struct then that's fine because the struct members because
those are handled in smatch_kernel_user_data.c already and nothing changes.
It's only when we start using p + 1 that this makes a difference.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>