param_set: fix param_was_set() for addresses
Say we have a parameter "p" and we set "p = foo;" then "&p->whatever;" is
not the same address as was passed in originally so the the
param_was_set_var_sym() function should return true.  But it wasn't set
up to handle addresses so it needed to be fixed.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>