param_key: fix some address handling
This patch got tangled with another thing I was working on and so it's
sort of a mess.  Somehow the key part of the param/key are getting
truncated in the database and I'm not sure where it happens.  It's probably
better to just not record anything instead of recording a truncated key.
The data is useless and not recording it will save disk space.  I added a
bunch of checks for snprintf() truncation and that's not the bug so I
stopped looking...

But the important part of this commit is that it improves how addresses are
handled.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>