scripts/kpatch: put a commented out diff into the commit message
When I'm writing the commit message then I always try to remember what the
patch actually does.  But you can't see it in the log, because it hasn't
been committed and you can't see it with `git diff` because it has been
added.

This patch is a game changer for writing commit messages.  It seem crazy
but that is really true for me.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>