locking: fix restore in cross function code
The bug here is when we have a function that accepts a parameter "flags"
and it does:

void my_function(..., unsigned long flags)
{
	...
	spin_unlock_irqrestore(lock, flags);
	...
	spin_lock_irqrestore(lock, flags);
}

The function was being recorded as a lock but actually the locks are the
same at the beginning and the end.  This change make restore match how
regular lock and unlock are handled.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>