missing_error_code: introduce new check
This is for in the kernel when there is a clean up block at the end of
the function.

	return 0;

cleanup_foo:
	clean_up(foo);
free_bar:
	kfree(bar);

	return ret;

The assumption is that this is an error path.  It's not always true.  The
check will not warn if there is a "ret = 0;" within three lines of the
goto so it won't warn about this:

		ret = 0;
		goto out;

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>