extra/ranges: introduce is_noderef_ptr()
There are are a few kinds of pointer values which are never supposed to
be dereferenced such as NULL, error pointer and poison pointers like
0xDEADBEEF.

There are some subtleties in how to handle them.  Say we have:

	foo->bar = 1;
	foo = <something>

If "<something> is a normal pointer then "foo->bar" is unknown, but if it
is a noderef pointer then we can't talk about foo->bar in a meaningful
sense.

I'm not sure exactly how to handle these.  But the first step is to add a
function so we can identify them.  (In other words this patch is slightly
useless by itself).

The one thing this patch does do is that it marks 0xdead pointers as "gone"
in the sense of parent_is_gone().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>