param_key: fix where assignments are handled
The problem is when we do:

	foo = param;

Then it gets translated into:

	foo = param;
	foo->member = fake();

So it looks like we're setting param->member to fake().  I had thought this
was fixed by moving the register_param_key in check_list.h until the end,
and possibly that fixed part of the problem but it doesn't work 100% for
sure.  The correct fix is to use ASSIGN_HOOK_LATE.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>