states: fix get_state_expr() uses of fake_parent
This has a couple problems I think...  Let's imagine a call:

	frob(foo);

The check might not track assignments...  Ideally it should but not
everything does.  Also it occurs to me that the parameter might be a
different type so it's not really accurate.

The original point of this was to track binops and function calls and I
still think that is a good idea...  We'll see.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>