fix possible circular definition with can_move_to()
can_move_to() is used to test if it is safe for a given pseudo to
be used by some instruction. This is done by checking that the
pseudo is defined 'before' the instruction.

This should, of course, reject the cases where the pseudo is defined
by the instruction itself because it would create a circular definition.
However, this special case was not checked.

Fix this by adding the missing check.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>