extra: fix test for take assignments
The problem is when we have a function call like "foo(x = whatever());"
The parameter gets changed into an assignment like __fake = x = whatever();
[ Revisit: if this is always necessary. ] so then both assignments are
treated as fake even though only one is.

The practical implications of this are that the equiv variables are not
tacked correctly.  And it messes up the parsing badly.  This is especially
important in the kernel because it affects the MIN() macro.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>