extra/function_hooks: handle fake function calls in extra instead of hooks
When we're doing struct assignments it gets translated into a bunch of
fake calls:

	foo->bar = unknown fake();

This was handled by smatch_function_hooks.c but it really should be handled
in smatch_extra.c since it's not a real function call.  The thing is that
these are really modifications per se and also we don't necessarily want to
create extra states for all of these.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>