function_hooks/check_unwind: add param_key_hooks
I really like how the check_locks.c and then later the check_unwind.c
info_tables work.  I think this way of writing checks is more flexible
and simpler.  The weird thing is that it heavily relies on the
return_implies hooks and for years I wanted to delete them but I just
never got around to it.  I'm calling this new style of check
"database first" style.

So what it means is that in the database you have something like
PARAM_FREE 0 "$".  Then the core will translate that into a var/sym pair
and pass that to the check.  This is simpler and more powerful than either
the old pre-database code where you had to get the argument pointer
yourself or the database code where you have to parse the param/key pair
yourself.  Of course, the idea that it will work the same for both
DB and table data, but I have not implemented the DB code yet.

This is just the first bit which affects the info table implementation.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>