fix add_join_conditional() when one of the alternative is VOID
add_join_conditional()'s job is to take the 2 alternatives
of the conditional, make a phi-node from them and return
the corresponding pseudo but if one of the alternatives
is not available it's useless to create a phi-node and
the other alternative can then directly be returned.

The problem is that in this later case, the pseudo directly
returned is the PSEUDO_PHI of the corresponding phi-source.
This gives erroneous code like, for example:
	phisrc.32	%phi1 <- $0
	ret.32		%phi1
instead of:
	ret.32		$0
since the %ph1 should only be used by a phi-node instruction.

Fix this by returning phi-source's operand instead.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>