db: fix how &foo->lock states are inserted into the caller_info table
The check_locks.c check has a lot of &foo->lock states and they're supposed
be inserted into the DB correctly when we call "frob(foo);".  But they
were not handled correctly by the caller_info callback (new style).  They
were handled but the CALL_HOOK instead which passes the parameter as -2.

Adding this missing case will fix that problem.  This code is sort of
copy and pasted and I'm not 100% comfortable that it's correct but it's
mostly correct.  Probably there is an easier and better what to manage
this stuff...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>