extra: fake assignments are also equivalent
I wrote this condition initially because I thought it would be a speed up.

The question is why are we creating fake assignments if the variables fits?

And the answer is "return foo ?: bar;"  So we do a fake assignment,
which sets gets changed to:  "return (__fake = foo) ?: __fake = bar;"
The select expression is too complicate so a fake assignment is required
but it gets translated into simple assignments so we need the equiv stuff
to be set up and to work so that the implications work.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>