extra: take a short cut in param_limit_binops
The db_param_limit_binops() function tries to say that if we know that
the $1 parameter is 0-9 range and we are passing "frob(i * 3);" then that
means "i is 0-3".

If the param limits "$->foo" then that's not useful for this so we can
just return early.  Hopefully it's a small speed up.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>