ptrlist: make ptr_list_nth_entry() generic with ptr_list_nth()
The library operation on pointer list necessarily act on the untyped
version of the list. To use them with type checking, they must either
be wrapped in inline function using the desired type or be used via
some macro doing the type checking.

Do this later solution for ptr_list_nth_entry().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>