locking: re-write and fixes
This patch is messier than I would wish but it just gets more and more
complicated to deal with it now.

The main thing is that I copied the format from check_unwind.c here.  It's
DB first style, to unify hard coded function calls with the DB calls.  So
the locking_table action becomes the same as for the DB.  The table now
includes the key where it didn't before.  Most of the time the key is "$"
but for IRQs the key is "irq" and that parameter is -2.  Also I added two
new functions "bch_write_bdev_super" where the key is "&$->sb_write_mutex"
so this added flexibility is already useful.

Instead of the "return_type" enum, the table now has the "implies_start/end"
information.  I'm not 100% sure this is an improvement, but cause if we
call spin_trylock() then before, the error path would be &unlocked and
now it's &start_state.  This is accurate enough, but it means that we miss
out on the double unlock warning.  Perhaps the correct thing is to make
the implies_start/end for the trylock failed set the state to unlocked.
(That thought just occured to me now).

Speaking of "double lock/unlock".  I've commented those warnings out for
now.  I'm still testing the code and recently made some changes so I'm not
ready to commit that yet.  The "called with lock held" warning still needs
work.  One problem is that are a bunch of file local locks with the same
name as each other so it generates false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>