param_set/clear: Use memset() of struct members
What this is is, imagine you have a function which memset as struct member:

void one(struct outer *outer)
{
        memset(&outer->foo, 0, sizeof(outer->foo));
}

That data is recorded in the database but we also record it as a series of
assignments.

	outer->foo.a = 0;
	outer->foo.b = 0;

It's a waste of space to record this duplicate information.  Originally
this param_clear was not used in smatch_struct_assignment.c but now we
can use the gen_expr_from_param_key() function to do it easily.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>