kvmalloc_NOFS: warn about passing bogus GFP_ arguments to kvmalloc()
The kvmalloc() function calls kmalloc() or if the memory is too much it
calls vmalloc().  But the vmalloc() function does not take a GFP_ argument.
Everything is just GFP_KERNEL.  So when you call kvmalloc() with GFP_ATOMIC
then it is equivalent to calling vanilla kmalloc(), the vmalloc() part is
not done.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>