flow: fake void returns if we have to bail on a function
For some reason, this function only faked a return for non-void functions.
The main reason for this function to exist is so that it records in the DB
that a function returns an unknown value so probably there isn't a reason
to record fake void returns...  But there is no reason not to either.

The other reason for this patch is that I wanted to use the
cur_func_return_type() as a clean up.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>