states: Rename delete_state() to __delete_state()
Back in the day, I used to try use delete_state() for flow analysis, but
these days what we do instead is set the state to &undefined.  The
delete_state() function actually doesn't work for flow analysis because if
you're in a condition then the deleted isn't copied to the final strees.

The function still serves a purpose for conserving memory.  There are some
places where we want to delete states just to conserve resources because we
know those states won't be used again.  It doesn't have a flow analysis
implication because we don't care what this states are.  And if we don't
delete everything that's fine too, so long as we delete some stuff then
that uses fewer resources so it's a savings.

Anyway, to mark this function as tricky and internal, then rename it from
delete_state() to __delete_state().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>