flow: pass conditional macros to MACRO_ASSIGNMENT_HOOK
I was writing a check which looks at min_t() assignments and none of them
are recorded.  Move the after_assign: label to take care of this.

There is still a "// FIXME: got after" and I tried to address it, but it
caused crashes so I left it as is.  This works for my min_t() check...
#Lazy

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>