flow: fake a return at the end of void functions
There are a bunch of places which want to track values when we return, for
example, the cross function DB or the locking check or the leak checks.
One thing is that they had to track returns and if end of the function if
the end of the function is reachable.  That's a little bit awkward and we
could avoid it by faking a return.

The __bail_on_rest_of_function case is ignored because it lead to problems.
Originally the code assumed that if the end of the function was reachable
then that meant it was a void function so it didn't have the check for
== &void_type.  But sometimes in a __bail_on_rest_of_function situation the
the end was reachable and it wasn't void.  I think the reason it was
reachable is because we were inside an if statement so the nullify path
didn't clear the main stree.  I'm not exactly sure what caused the crash
but maybe faking a void return in a non-void function is a problem.

In theory just checking for __bail is sufficient to avoid the crash but
the == &void_ctype check adds readability and robustness.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>