points_to_user_data: fix is_skb_data()
Some of these changes are just stuff that I did because they looked right.
But the most important change is to just use get_type() to get the type.
Before if you had code like "foo->skb->data" then it would use the type
of "foo" which is not going to be an sk_buff.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>