points_to_user_data: handle memcpy() correctly
The original code had a weird thing where it said that memcpy() results
in a fake assignment and hacked around that.  In my test case, it seemed to
not work correctly.

I didn't investigate why, but instead just re-wrote this so it looks how
you would expect with an explicit hook for memcpy().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>