function_hooks/user_data: introduce and use add_function_hook_late()
The truth is that when I look at this code, I don't know how the kstrtol()
functions ever worked correctly but when I look at my database I can see
they were working until two days ago.  I'm not sure what happened.
Perhaps the changes to how dereferences are parsed exposed this bug.

Anyway the problem is the ordering is totally wrong.  First we set the data
to user data.  Then we parse the information from the DB and it's a
PARAM_SET so the data is modified so now it's not user data any more.
Except on the error paths where the variable is maybe uninitialized still
and also user data.

The other interesting thing is that there is a check in the kstrotoint()
and similar functions like this:

	if (result != (int)result)
		return -ERANGE;

Smatch sees on the success path that "result == result" and decides that
makes it capped.  Ugh...

Anyway, the fix for all these is to set it to user data after the database
information has been parsed.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>