mtag_data: store mtag_data for structs declared on the stack
The problem is that if we have an mtag then we really need to record the
data.  The probably is we pass the mtag to another function and then that
sets something so that gets recorded but not the local stuff.  So then
there is a kind of mismatch.

Why even give local structs an mtag value if we're not going to use it?

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>